Fix compilation warnings.

+++ particle/flameparticle.cpp
@@ -81,7 +81,8 @@ void CFlameParticle::exit()
 
 void CFlameParticle::draw()
 {
-	g->DrawFrameClip(frame, pos.x - g->Width / 2, pos.y - g->Height / 2);
+	g->DrawFrameClip(frame, static_cast<int>(pos.x - g->Width / 2),
+		static_cast<int>(pos.y - g->Height / 2));
 }
 
 void CFlameParticle::update(int ticks)
